+++ tests/Justimmo/Tests/Model/RealtyQueryTest.php
@@ -14,6 +14,7 @@ class RealtyQueryTest extends TestCase
     private function getQuery($api = null)
     {
         $mapper = new RealtyMapper();
+
         return new RealtyQuery(($api ?: new JustimmoNullApi()), new RealtyWrapper($mapper), $mapper);
     }
 
@@ -28,9 +29,47 @@ class RealtyQueryTest extends TestCase
         $this->assertEquals(array('alleProjektObjekte' => 0), $query->getParams());
     }
 
+    public function testFilterByUpdatedAt()
+    {
+        $query = $this->getQuery();
+        $query->filterByUpdatedAt('2017-01-12');
+        $this->assertEquals(array(
+            'filter' => array(
+                'aktualisiert_am' => '2017-01-12',
+            ),
+        ), $query->getParams());
+
+        $query = $this->getQuery();
+        $query->filterByUpdatedAt(array('min' => '2017-01-12', 'max' => '2017-01-15'));
+        $this->assertEquals(array(
+            'filter' => array(
+                'aktualisiert_am_von' => '2017-01-12',
+                'aktualisiert_am_bis' => '2017-01-15',
+            ),
+        ), $query->getParams());
+    }
+
+    public function testOrderByUpdatedAt()
+    {
+        $query = $this->getQuery();
+        $query->orderByUpdatedAt('asc');
+
+        $this->assertEquals(array(
+            'orderby'   => 'updated_at',
+            'ordertype' => 'asc',
+        ), $query->getParams());
+
+        $query->orderByUpdatedAt('desc');
+
+        $this->assertEquals(array(
+            'orderby'   => 'updated_at',
+            'ordertype' => 'desc',
+        ), $query->getParams());
+    }
+
     public function testFindIds()
     {
-        $api = new MockJustimmoApi(array('realtyIds' => $this->getFixtures('v1/realty_ids.json')));
+        $api   = new MockJustimmoApi(array('realtyIds' => $this->getFixtures('v1/realty_ids.json')));
         $query = $this->getQuery($api);
 
         $this->assertEquals(array(
@@ -38,7 +77,7 @@ class RealtyQueryTest extends TestCase
             728584,
             635873,
             587626,
-            587622
+            587622,
         ), $query->findIds());
     }
 }