+++ CHANGELOG.md
@@ -2,6 +2,7 @@
  * Add support for project state, url, sale start, completion date and is reference in project calls
  * Add support for plain rent fields without additional costs (rent net, rent gross, rent vat type, rent vat value, rent vat input and rent vat) in realty calls
  * Add support for allProjectRealties setting
+ * Add support for filterByUpdatedAt in realty calls
  * Add support for employee ids and project ids call
 
 ## 1.1.2